Added explosions and a way for the missile to deal AOE damage

+++ Assets/Scripts/Interfaces/ITakeDamagable.cs
@@ -1,5 +1,5 @@
 ﻿public interface ITakeDamagable
 {
-    float Health { get; }
+    int Health { get; }
     void TakeDamage(int damage);
 }
